build: Use scalac as our default compiler to be able to compile scala code
This is step 1 of the Scala integration. We now use the scalac task
to compile our code. We also integrate the scalatest classes that can be used to
generated class files that can be run using JUnit. The fsc task is for faster
compilation as it uses a daemon to reduce startup time. It is not the default
compile mode as it can mess with flags.

Signed-off-by: David Soria Parra <dsp@php.net>