Calculator: Use a Scala implementation of the Calculator
As a first step to refactor the current calculator code, we remove
the Calculator interface and replace it with an abstract class
written in Scala. The abstract class will offer additional methods
that can be used to generate graphs. In a later step we'll add a
set of traits that can activly modify these methods to enhance the
generation methods.

We also have to adept the interface little bit to use the Integer class.

Please note that we will write Scala 2.8 code and rely on the new
collections API.

Signed-off-by: David Soria Parra <dsp@php.net>