Fix a possible NodeDoesNotExist exception
We might run into a possible node does not exist exception if we try to
get a node that is not in the original graph. We have to avoid this. We also
need to unset the central node in the global calculator. This behavior can lead
to a possible race condition between peer and global calculator and is tracked
by issue #1.

Related-to: Issue #1

Signed-off-by: David Soria Parra <dsp@php.net>