Make sure the URL class loader is assigned to be the current thread classloader
We have to make sure that the URL class loader that we use is also the current
thread classloader as this is the one that for example Clojure will use to find
the .clj files in the .jar file. This fix was suggested by Stuard Halloway in
ticket #260 of the clojure ticket system.

Signed-off-by: David Soria Parra <dsp@php.net>